LayoutTests/css3/scroll-snap/ios/scroll-snap-mainframe-pinch-out-expected.txt

This test verifies that pinching out below the minimum scale in a root scroll
        snapping container while triggering layout passes does not cause us to snap to an incorrect
        position upon finishing the zoom animation. To manually test, pinch out to zoom on this page
        and verify that the final scroll position is the same as the initial position (marked 4),
        and the scroll position does not flicker in the process.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS finalOffsetY is originalOffsetY
PASS successfullyParsed is true

TEST COMPLETE
